size_t has different widths on different platforms and %z isn't portable so use %d and cast to int. (st)
git-svn-id: svn://mielke.cc/main/brltty@1910 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ BrailleDrivers/Braudi/braille.c
@@ -152,7 +152,7 @@ identifyDisplay (BrailleDisplay *brl) {
             skipCharacter(' ', &bytes, &count);
             if (interpretNumber(&cells, &bytes, &count)) {
               if (!count) {
-                LogPrint(LOG_INFO, "Detected: %.*s", length, identity);
+                LogPrint(LOG_INFO, "Detected: %.*s", (int)length, identity);
 
                 brl->x = cells;
                 brl->y = 1;
